server/requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
requests==2.32.3
smmap==5.0.1
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
